(add) lucem: multi-user support
Lucem no longer outputs the Sober log to a centralized location, rather
it outputs it to a user's `XDG_RUNTIME_DIR` whenever possible, allowing
for multiple users to play with Lucem on a single machine without
conflicting with each others' instances. This is a pretty useless
feature and niche as hell, but it's nice nonetheless.